Last Time A* is Fixed
The problem with A* was that I added everything into the Heap before my
F value was calculated.  This would have been ok if the Heap wasn't
sorted (like a list) and it was sorted as we removed elements.  But
since it gets sorted when the element is added it already needs to have
calculated it's F value