Temporary menu and "pause menu" implemented.
Added a menu that is mostly just a scene navigator. The pause menu can
be accessed in the Main scene by pressing Escape, which puts up an
overlay and doesn't pause the game (I'll look into pausing local games
but not network games).

The exit option in the menus only closes the game if you build it, so if
you try using it in Unity's viewer, it won't work.